FIX:Nutrientのデータ型修正 FIX:show_recipeのレイアウト修正 FIX:dietary_fiberとsalt_equivalentを小数点第1位で四捨五入 FIX:スクレイピングの変数の初期値

+++ db/migrate/20220409125541_create_nutrients.rb
@@ -6,8 +6,8 @@ class CreateNutrients < ActiveRecord::Migration[6.1]
       t.integer :carbohydrate
       t.integer :protein
       t.integer :lipid
-      t.integer :dietary_fiber
-      t.integer :salt_equivalent
+      t.float :dietary_fiber
+      t.float :salt_equivalent
 
       t.timestamps
     end
